GPBN-2 -Add admin user auto adding in the begging
