+++ gpb-web/src/main/resources/application.properties
@@ -5,8 +5,6 @@ spring.datasource.hikari.schema = public
 spring.datasource.driverClassName = org.postgresql.Driver
 spring.jpa.generate-ddl = true
 
-spring.flyway.schemas=public
-spring.flyway.locations=classpath:db/migration
 
 # Hibernate
 hibernate.dialect: org.hibernate.dialect.PostgreSQLDialect
@@ -14,6 +12,8 @@ hibernate.show_sql: true
 hibernate.hbm2ddl.auto: update
 
 spring.resources.add-mappings=true
+
+spring.flyway.enabled=false
 
 logging.level.org.springframework.security=INFO
 logging.level.org.apache.kafka=INFO
