+++ gpb-web/src/test/java/com/gpb/web/integration/game/BaseAuthenticationIntegration.java
@@ -109,6 +109,8 @@ public class BaseAuthenticationIntegration {
         games.add(gameCreation("name1", "url1", Genre.STRATEGIES, new BigDecimal("100.0"), new BigDecimal("100.0")));
         games.add(gameCreation("name2", "url2", Genre.RPG, new BigDecimal("500.0"), new BigDecimal("500.0")));
         games.add(gameCreation("name3", "url3", Genre.STRATEGIES, new BigDecimal("1000.0"), new BigDecimal("800.0")));
+        System.setProperty("ADMIN_EMAIL", "");
+        System.setProperty("ADMIN_PASSWORD_HASH", "");
         System.setProperty("GAMEZEY_LOGIN", "");
         System.setProperty("GAMEZEY_PASSWORD", "");
         System.setProperty("IMAGE_FOLDER", "");
